Changed files
Changed folder name to gui, updated some styles and removed some JQuery
