--- TSE/static/js/custom.js
@@ -1,20 +1,0 @@
-$(document).ready(function() {
-  // Prototype one buttons
-  $("#uni-btn").click(function() {
-    document.location.href = "questions-p1.html";
-  });
-  $("#det-btn").click(function() {
-    document.location.href = "questions-p1.html";
-  });
-  $("#acc-btn").click(function() {
-    document.location.href = "questions-p1.html";
-  });
-  $("#oth-btn").click(function() {
-    document.location.href = "questions-p1.html";
-  });
-
-  // Core buttons
-  $("#search-btn").click(function() {
-    // Search through database
-  });
-});
