+++ gui/static/css/styles.css
@@ -2,6 +2,9 @@
 body {
   background-color: #efefef !important;
 }
+a {
+  color: inherit !important;
+}
 .spacing {
   padding-left: 20px;
   padding-right: 20px;
@@ -140,21 +143,20 @@ body {
 .category.first {
   background-color: #286090;
 }
-.category.first button {
-  color: #fff;
+.category.first a#uni-btn {
+  color: #fff !important;
 }
-.category button {
+.category a {
   background-color: inherit;
   border: none;
   color: #000;
   font-family: myriad-pro, sans-serif;
   font-size: 48px;
+  line-height: 28.5vh;
+  display: block;
   text-align: center;
   width: 100%;
   height: 100%;
-}
-.category button {
-  color: #000;
 }
 .p-one.return-btn {
   margin-top: 10px;
