+++ gui/static/js/custom.js
@@ -1,0 +1,6 @@
+$(document).ready(function() {
+  // Core buttons
+  $("#search-btn").click(function() {
+    // Search through database
+  });
+});
