+++ gui/templates/home-p1.html
@@ -26,19 +26,19 @@
 
     <div class="row vertical-spacing justify-content-center">
       <div class="col-lg-5 mx-4 category first">
-        <button id="uni-btn" type="button"><a href="{{ url_for('questions') }}">University</a></button>
+        <a id="uni-btn" href="{{ url_for('questions') }}">University</a>
       </div><!-- /.category -->
       <div class="col-lg-5 mx-4 category">
-        <button id="det-btn" type="button"><a href="{{ url_for('questions') }}">Course Details</a></button>
+        <a id="det-btn" href="{{ url_for('questions') }}">Course Details</a>
       </div><!-- /.category -->
     </div><!-- /.row -->
 
     <div class="row vertical-spacing justify-content-center">
       <div class="col-lg-5 mx-4 category">
-        <button id="acc-btn" type="button"><a href="{{ url_for('questions') }}">Accommodation</a></button>
+        <a id="acc-btn" href="{{ url_for('questions') }}">Accommodation</a>
       </div><!-- /.category -->
       <div class="col-lg-5 mx-4 category">
-        <button id="oth-btn" type="button"><a href="{{ url_for('questions') }}">Other</a></button>
+        <a id="oth-btn" href="{{ url_for('questions') }}">Other</a>
       </div><!-- /.category -->
     </div><!-- /.row -->
   </div> <!-- /.container -->
